pyproject.toml
[project]
name = "Palladia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diff-match-patch>=20241021",
    "jiwer>=4.0.0",
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "openrouter>=0.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
